# Benchmark and further optimize spatialmatch.stackable()