--- a/lib/spatialmatch.js
+++ b/lib/spatialmatch.js
@@ -163,63 +163,86 @@ function stackByIdx(stack) {
 // For a given set of phrasematch results across multiple indexes,
 // provide all relevant stacking combinations using phrase masks to
 // exclude colliding matches.
-function stackable(stacks, phrasematches, idx, mask, nmask, stack, relev) {
-    idx = idx || 0;
-    mask = mask || 0;
-    nmask = nmask || 0;
-    stack = stack || [];
-    relev = relev || 0;
-
-    if (!phrasematches[idx]) return stacks;
+function stackable(stacks, phrasematches, limit, max, idx, mask, nmask, stack, relev) {
+    if (idx === undefined) {
+        idx = 0;
+        limit = 30;
+        mask = 0;
+        nmask = 0;
+        stack = [];
+        relev = 0;
+        max = { stacks:[], relev:0, count:0, calls: 0 };
+    }
 
     // Recurse, skipping this level
-    stackable(stacks, phrasematches, idx+1, mask, nmask, stack, relev);
+    if (phrasematches[idx+1] !== undefined) {
+        stackable(stacks, phrasematches, limit, max, idx+1, mask, nmask, stack, relev);
+    }
 
     // For each stacked item check the next bmask for its idx.
     // If the bmask includes the idx these indexes cannot stack
     // (their geocoder_stack do not intersect at all).
     var bmask = phrasematches[idx].length && phrasematches[idx][0].bmask;
     if (bmask) for (var j = 0; j < stack.length; j++) {
-        if (bmask[stack[j].idx]) return stacks;
+        if (bmask[stack[j].idx]) return;
     }
 
     // Recurse, including this level
     for (var i = 0; i < phrasematches[idx].length; i++) {
-        var targetMask = mask;
-        var targetNmask = nmask;
         var next = phrasematches[idx][i];
-        var direction;
+        if (mask & next.mask) continue;
+        if (nmask & next.nmask) continue;
 
-        if (targetMask & next.mask) continue;
-        if (targetNmask & next.nmask) continue;
-
-        // sort by order of input query
-        var targetStack = stack.slice(0);
-        targetStack.relev = relev;
-        targetStack.sort(sortByMask);
         // compare index order to input order to determine direction
-        if (targetStack.length) {
-            direction = targetStack[0].idx < next.idx ? "descending" : "ascending";
+        if (stack.length &&
+            stack[0].idx >= next.idx &&
+            mask &&
+            mask < next.mask) continue;
+
+        var targetStack = stack.slice(0);
+        var targetMask = mask | next.mask
+        var targetNmask = nmask | next.nmask;
+        targetStack.relev = relev + next.weight;
+
+        // ensure order of targetStack maintains lowest mask value at the
+        // first position. ensure direction check above works.
+        if (next.mask < mask) {
+            targetStack.unshift(next);
+        } else {
+            targetStack.push(next);
         }
-
-        if (direction === "ascending" && targetMask && targetMask < next.mask) continue;
-
-        targetMask = targetMask | next.mask;
-        targetNmask = targetNmask | next.nmask;
-        targetStack.push(next);
-        targetStack.relev += next.weight;
 
         if (targetStack.relev > 0.5) {
-            targetStack.sort(sortByZoomIdx);
-            stacks.push(targetStack);
+            if (targetStack.relev > max.relev) {
+                max.count = 1;
+                max.relev = targetStack.relev;
+                max.stacks = [ targetStack ];
+                stacks.push(targetStack);
+            } else if (targetStack.relev === max.relev) {
+                max.count++;
+                max.stacks.push(targetStack);
+                if (max.count < limit) {
+                    stacks.push(targetStack);
+                } else if (max.count === limit) {
+                    stacks = max.stacks;
+                }
+            } else if (max.count < limit) {
+                stacks.push(targetStack);
+            }
         }
 
-        stackable(stacks, phrasematches, idx+1, targetMask, targetNmask, targetStack, targetStack.relev);
+        // Recurse to next level
+        if (phrasematches[idx+1] !== undefined) {
+            stackable(stacks, phrasematches, limit, max, idx+1, targetMask, targetNmask, targetStack, targetStack.relev);
+        }
     }
 
-    if (idx === 0) stacks.sort(sortByRelevLengthIdx);
-
-    return stacks;
+    if (idx === 0) {
+        stacks.forEach(function(stack) { stack.sort(sortByZoomIdx); });
+        stacks.sort(sortByRelevLengthIdx);
+        stacks = stacks.slice(0, limit);
+        return stacks;
+    }
 }
 
 function sortByRelevLengthIdx(a, b) {
@@ -230,19 +253,13 @@ function sortByRelevLengthIdx(a, b) {
 }
 
 function sortByZoomIdx(a, b) {
-    return (a.zoom - b.zoom) || (b.idx - a.idx);
+    return (a.zoom - b.zoom) || (b.idx - a.idx) || (b.mask - a.mask);
 }
 
 function sortByRelev(a, b) {
     return (b.relev - a.relev) ||
         (b[0].scoredist - a[0].scoredist) ||
         (a[0].idx - b[0].idx);
-}
-
-function sortByMask(a, b) {
-    if (!b) return 1;
-    if (!a) return 0;
-    return (a.mask - b.mask);
 }
 
 function rebalance(query, stack) {