--- a/lib/spatialmatch.js
+++ b/lib/spatialmatch.js
@@ -9,7 +9,7 @@ module.exports.stackable = stackable;
 module.exports.rebalance = rebalance;
 
 function spatialmatch(query, phrasematches, options, callback) {
-    var stacks = stackable([], phrasematches);
+    var stacks = phrasematches.length ? stackable(phrasematches) : [];
     stacks = allowed(stacks, options);
     stacks = stacks.slice(0,30);
     // Rebalance weights, relevs of stacks here.
@@ -163,20 +163,23 @@ function stackByIdx(stack) {
 // For a given set of phrasematch results across multiple indexes,
 // provide all relevant stacking combinations using phrase masks to
 // exclude colliding matches.
-function stackable(stacks, phrasematches, limit, max, idx, mask, nmask, stack, relev) {
-    if (idx === undefined) {
+function stackable(phrasematches, memo, idx, mask, nmask, stack, relev) {
+    if (memo === undefined) {
+        memo = {
+            stacks: [],
+            maxStacks: [],
+            maxRelev: 0
+        };
         idx = 0;
-        limit = 30;
         mask = 0;
         nmask = 0;
         stack = [];
         relev = 0;
-        max = { stacks:[], relev:0, count:0, calls: 0 };
     }
 
     // Recurse, skipping this level
     if (phrasematches[idx+1] !== undefined) {
-        stackable(stacks, phrasematches, limit, max, idx+1, mask, nmask, stack, relev);
+        stackable(phrasematches, memo, idx+1, mask, nmask, stack, relev);
     }
 
     // For each stacked item check the next bmask for its idx.
@@ -186,6 +189,10 @@ function stackable(stacks, phrasematches, limit, max, idx, mask, nmask, stack, r
     if (bmask) for (var j = 0; j < stack.length; j++) {
         if (bmask[stack[j].idx]) return;
     }
+
+    // Though spatialmatches are sliced to 30 elements after this step
+    // leave some headroom as this step does not include type filtering.
+    var limit = 100;
 
     // Recurse, including this level
     for (var i = 0; i < phrasematches[idx].length; i++) {
@@ -213,43 +220,46 @@ function stackable(stacks, phrasematches, limit, max, idx, mask, nmask, stack, r
         }
 
         if (targetStack.relev > 0.5) {
-            if (targetStack.relev > max.relev) {
-                max.count = 1;
-                max.relev = targetStack.relev;
-                max.stacks = [ targetStack ];
-                stacks.push(targetStack);
-            } else if (targetStack.relev === max.relev) {
-                max.count++;
-                max.stacks.push(targetStack);
-                if (max.count < limit) {
-                    stacks.push(targetStack);
-                } else if (max.count === limit) {
-                    stacks = max.stacks;
+            if (targetStack.relev > memo.maxRelev) {
+                if (memo.maxStacks.length >= limit) {
+                    memo.stacks = memo.maxStacks;
+                    memo.maxStacks = [ targetStack ];
+                } else {
+                    memo.maxStacks.push(targetStack);
                 }
-            } else if (max.count < limit) {
-                stacks.push(targetStack);
+                memo.maxRelev = targetStack.relev;
+            } else if (targetStack.relev === memo.maxRelev) {
+                memo.maxStacks.push(targetStack);
+            } else if (memo.maxStacks.length < limit) {
+                memo.stacks.push(targetStack);
             }
         }
 
         // Recurse to next level
         if (phrasematches[idx+1] !== undefined) {
-            stackable(stacks, phrasematches, limit, max, idx+1, targetMask, targetNmask, targetStack, targetStack.relev);
+            stackable(phrasematches, memo, idx+1, targetMask, targetNmask, targetStack, targetStack.relev);
         }
     }
 
     if (idx === 0) {
+        var stacks;
+        stacks = memo.stacks.concat(memo.maxStacks);
         stacks.forEach(function(stack) { stack.sort(sortByZoomIdx); });
         stacks.sort(sortByRelevLengthIdx);
-        stacks = stacks.slice(0, limit);
-        return stacks;
+        return stacks.slice(0, limit);
     }
 }
 
 function sortByRelevLengthIdx(a, b) {
-    return (b.relev - a.relev) ||
+    var first = (b.relev - a.relev) ||
         (a.length - b.length) ||
-        (b[b.length-1].scorefactor - a[a.length-1].scorefactor) ||
-        (a[a.length-1].idx - b[b.length-1].idx);
+        (b[b.length-1].scorefactor - a[a.length-1].scorefactor);
+    if (first) return first;
+
+    for (var end = a.length - 1; end >= 0; end--) {
+        var second = a[end].idx - b[end].idx;
+        if (second) return second;
+    }
 }
 
 function sortByZoomIdx(a, b) {