### Background: what does spatialmatch.stackable() do?

For a given set of phrasematches, `spatialmatch.stackable()` generates all the possible/allowed stacking permutations that we should consider as potential geocoding matches in the real world.

**Simplified example:** If we matched

```
100 main street (poi)
100 main street (address)
new york (city)
new york (state)
```

Stackable says we need to check for:

```
100 main street (poi)
100 main street (address)
new york (city)
new york (state)
100 main street (poi), new york (city)
100 main street (poi), new york (state)
100 main street (address), new york (city)
100 main street (address), new york (state)
```

Then would sort these in order of most likely relevance.

------

### What are we trying to improve?

With long queries that match many possible phrases across a lot of indexes, this list of permutation grows exponentially and takes a very long time to generate.

**Goal:** make it faster, and make it not crash under extreme conditions.

- **Adds synthetic stress tests for `spatialmatch.stackable()`.** These tests are pretty unrealistic but telling -- they simulate many indexes matching many possible combinations of phrases and then time how quickly stackable can generate permutations for them.
- **Optimizes `spatialmatch.stackable()`.** Because we use only the top 30 permutations, sorted by relev, the biggest strategy here is to store the maximum relev we've seen as we generate permutations and strategically throw out permutations as we go that don't meet the maximum relev threshold. A bunch of other micro JS optimizations tune things further but are more tweaks.

------

### Results

Test | Before | After
--- | --- | ---
5 indexes x 10 terms | 23ms | 17ms
6 indexes x 10 terms | 92ms | 33ms
7 indexes x 10 terms | 551ms | 90ms
8 indexes x 10 terms | 1631ms | 225ms
9 indexes x 10 terms | 7098ms | 360ms
10 indexes x 10 terms | n/a* | 8297ms

n/a - under certain circumstances this testrun on `master` will crash, likely from running out of memory.

```
<--- Last few GCs --->

  192174 ms: Mark-sweep 1399.5 (1457.3) -> 1399.5 (1457.3) MB, 2488.6 / 0 ms [last resort gc].
  192177 ms: Scavenge 1399.5 (1457.3) -> 1399.5 (1457.3) MB, 3.1 / 0 ms [allocation failure].
  192181 ms: Scavenge 1399.5 (1457.3) -> 1399.5 (1457.3) MB, 2.8 / 0 ms [allocation failure].
  194626 ms: Mark-sweep 1399.5 (1457.3) -> 1399.5 (1457.3) MB, 2444.6 / 0 ms [last resort gc].
  197077 ms: Mark-sweep 1399.5 (1457.3) -> 1399.5 (1457.3) MB, 2451.7 / 0 ms [last resort gc].


<--- JS stacktrace --->

==== JS stack trace =========================================

Security context: 0x2a0e013b4629 <JS Object>
    2: InnerArraySort [native array.js:~670] [pc=0x29130511e4b7] (this=0x3439b8004231 <JS Array[5]>,v=5,aC=0x1e71f2abad31 <JS Function sortByMask (SharedFunctionInfo 0x1e71f2a3d939)>)
    3: sort [native array.js:~885] [pc=0x291304e3d822] (this=0x3439b8004231 <JS Array[5]>,aC=0x1e71f2abad31 <JS Function sortByMask (SharedFunctionInfo 0x1e71f2a3d939)>)
    4: stackable(aka stackable) [/home/mapb...

FATAL ERROR: CALL_AND_RETRY_LAST Allocation failed - process out of memory
Aborted (core dumped)
```

------

- [x] Integration/staging testrun
- [x] Roll new release

cc @mapbox/geocoding-gang @ianshward